this commit was sponsored by double-precision gang

+++ include/ray.h
@@ -8,11 +8,11 @@ struct Ray {
     Ray() {};
     Ray(const Vec3& origin, const Vec3& direction) : origin(origin), direction(direction) {}
 
-    inline Vec3 PositionAt(float t) const { return origin + t*direction; }
+    inline Vec3 PositionAt(double t) const { return origin + t*direction; }
 
     // calculate reflecting ray given the position and normal
     Ray Reflect(const Vec3& position, const Vec3& normal) const;
-    Ray Refract(const Vec3& position, const Vec3& normal, float n1, float n2) const;
+    Ray Refract(const Vec3& position, const Vec3& normal, double n1, double n2) const;
 };
 
 #endif
