+++ src/ray.cpp
@@ -2,14 +2,14 @@
 
 static inline Vec3 ReflectVec(const Vec3& v, const Vec3& n)
 {
-    return v - 2.f*Dot(v, n)*n;
+    return v - 2*Dot(v, n)*n;
 }
 
-static inline Vec3 RefractVec(const Vec3& v, const Vec3& n, float ratio)
+static inline Vec3 RefractVec(const Vec3& v, const Vec3& n, double ratio)
 {
-    float dt = Dot(v, n);
-    float discriminant = 1.f - ratio*ratio*(1.f - dt*dt);
-    if(discriminant <= 0.f) return {};
+    double dt = Dot(v, n);
+    double discriminant = 1.0 - ratio*ratio*(1.0 - dt*dt);
+    if(discriminant <= 0.0) return {};
     return ratio*(v - n*dt) - n*sqrtf(discriminant);
 }
 
@@ -19,9 +19,9 @@ Ray Ray::Reflect(const Vec3& position, const Vec3& normal) const
     return { position, reflect_dir };
 }
 
-Ray Ray::Refract(const Vec3& position, const Vec3& normal, float n1, float n2) const
+Ray Ray::Refract(const Vec3& position, const Vec3& normal, double n1, double n2) const
 {
-    float ratio = n1 / n2;
+    double ratio = n1 / n2;
     Vec3 refract_dir = RefractVec(this->direction, normal, ratio);
     return { position, refract_dir };
 }
